clean PythonWorksp/TensorFlow
