+++ PythonWorksp/TensorFlow/MNIST/ckpt/checkpoint
@@ -1,0 +1,2 @@
+model_checkpoint_path: "newpath-2000"
+all_model_checkpoint_paths: "newpath-2000"
